Move isSiteBlocked to state/reader
