+++ client/reader/feed-stream/index.jsx
@@ -19,7 +19,7 @@ import QueryReaderFeed from 'components/data/query-reader-feed';
 import { getSite } from 'state/reader/sites/selectors';
 import { getFeed } from 'state/reader/feeds/selectors';
 import { getSiteName } from 'reader/get-helpers';
-import isSiteBlocked from 'state/selectors/is-site-blocked';
+import isSiteBlocked from 'state/reader/site-blocks/selectors/is-site-blocked';
 import SiteBlocked from 'reader/site-blocked';
 
 // If the blog_ID of a reader feed is 0, that means no site exists for it.
